requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
audioop-lts==0.2.2
brotli==1.2.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
fastapi==0.136.1
filelock==3.29.0
flatbuffers==25.12.19
fsspec==2026.3.0
gradio==6.13.0
gradio_client==2.5.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.4.1
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
ImageIO==2.37.3
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-loader==0.5
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
onnxruntime==1.25.1
orjson==3.11.8
packaging==26.2
pandas==3.0.2
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
protobuf==7.34.1
pydantic==2.13.3
pydantic_core==2.46.3
pydub==0.25.1
Pygments==2.20.0
PyMatting==1.1.15
python-dateutil==2.9.0.post0
python-multipart==0.0.27
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.75
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
safehttpx==0.1.7
scikit-image==0.26.0
scipy==1.17.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
tifffile==2026.4.11
tomlkit==0.14.0
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
